+++ main/sensors.h
@@ -13,7 +13,7 @@
  *   pH = V_sensor * 2.8
  *   Temperature = V_sensor * 20.0 - 20.0
  *
- * @author  Biofloc Engineering Team
+ * @author  @Marton1123 (https://github.com/Marton1123)
  * @date    2026
  */